get_summary move to presenter
